backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
openai==1.30.0
feedparser==6.0.11
httpx==0.27.0
apscheduler==3.10.4
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-multipart==0.0.9
pytz==2024.1
# --- 新增：公开研究平台 ---
psycopg2-binary==2.9.9
pypinyin==0.52.0
python-slugify==8.0.4
